// generated/src/app/api-models/claim-productivity-trans.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimProductivityTrans {
  seqTransId: number;
  transDateTime: Date;
  userId: string;
  claimType: string;
  primaryDate: Date;
  seqClaimId: number;
  lineNumber: number;
  subLineCode: string;
  transType: string;
  receivedDate: Date;
  batchNumber: string;
  adjudicationMethod: string;
  seqGroupId: number;
  lineOfBusiness: string;
  originalUserId: string;
  claimStatus: string;
}